requirements.txt
phidata
python-dotenv
openai
duckduckgo-search
yfinance
exa-py
lancedb 
tantivy 
pypdf 
sqlalchemy
fastapi
python-multipart
uvicorn
streamlit
pgvector
psycopg
typer
psycopg2_binary
youtube_transcript_api
PyPDF2
duckdb